backend/requirements.txt
pydantic==2.9.2
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.5.0
python-dotenv==1.0.0
openai==1.3.0
# Use a wheel that doesn't require local C build tools.
# For Python 3.13 on Windows, pin a newer pandas with prebuilt wheels.
pandas==2.3.3
openpyxl==3.1.2
# PostgreSQL dependencies - UPDATED for Python 3.13 compatibility
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
alembic==1.13.1